Change tx_count size in Block to u32

+++ exonum/src/api/mod.rs
@@ -227,7 +227,7 @@ mod tests {
     fn test_json_response_for_complex_val() {
         let str_val = "sghdkgskgskldghshgsd";
         let txs = [34, 32];
-        let tx_count = txs.len() as u64;
+        let tx_count = txs.len() as u32;
         let complex_val = Block::new(SCHEMA_MAJOR_VERSION,
                                      0,
                                      24,
